+++ app_feup/lib/main.dart
@@ -13,7 +13,7 @@ import 'package:flutter_redux/flutter_redux.dart';
 import 'view/Theme.dart';
 import 'model/AppState.dart';
 import 'package:redux/redux.dart';
-import 'redux/reducers.dart';
+import 'redux/Reducers.dart';
 import 'controller/parsers/parser-exams.dart';
 import 'package:app_feup/controller/LifecycleEventHandler.dart';
