+++ app_feup/lib/view/widgets/NavigationDrawer.dart
@@ -2,7 +2,7 @@ import 'package:app_feup/view/Theme.dart';
 import 'package:flutter/material.dart';
 
 import 'package:flutter_redux/flutter_redux.dart';
-import 'package:app_feup/redux/actionCreators.dart';
+import 'package:app_feup/redux/ActionCreators.dart';
 import 'package:app_feup/model/AppState.dart';
 
 class NavigationDrawer extends StatefulWidget {
